web-app/requirements.txt
ffmpeg==1.4
Flask==3.0.0
numpy==1.24.4
openai-whisper==20231117
pylint==3.0.2
pymongo==4.6.0
regex==2023.10.3
requests==2.31.0
torch==2.1.1
tqdm==4.66.1
urllib3==2.1.0
Werkzeug==3.0.1
